modules/stdpass.fpd: added experimental -M variants for manual soldering that have 0.05 mm longer pads.
The regular footprints for 0402 and 0603 are designed for reflow soldering
and they should work well for that. However, when soldering with a
soldering iron, often a little more space on the short side of the
component would be desirable. This change provides a set of components
that add 0.05 mm to each pad, making the entire footprint (and outline)
0.1 mm longer.

This is exerimental and may change in the future.