modules/meander-2450MHz.fpd (-RELAXED): grow bottom end of the antenna by 0.1 mm
... for a more relaxed fit of the ground via.

Background: if using a 0.3 mm drill for a finished hole size of 0.2 mm,
the drill to pad edge clearance is only 0.1 mm on the top and bottom of
the pad. To relax the constraints on the manufacturing process, we grow
the pad a little, to add an extra 2 * 0.05 mm of clearance.